refactor(studio): share spread-pair normalization and fix extract-stage range filter
Extract normalizeLeads into a shared spread-picker/pairs helper (the single
source of truth mirroring the extractor's greedy grouping) and use it from the
editor, dialog, wizard card, and extract stage. Fixes the extract stage
counting/showing marked spreads outside the current page range. Trim redundant
comments.
